+++ src/main/java/net/hexogendev/hexogen/api/world/generator/utils/NoiseGeneratorPerlin.java
@@ -36,7 +36,7 @@ public class NoiseGeneratorPerlin extends NoiseGenerator {
 
 	public final double func_4110_a(int i, double d, double d1) {
 		int j = i & 0xf;
-		double d2 = (double) (1 - ((j & 8) >> 3)) * d;
+		double d2 = (1 - ((j & 8) >> 3)) * d;
 		double d3 = j >= 4 ? j != 12 && j != 14 ? d1 : d : 0.0D;
 		return ((j & 1) != 0 ? -d2 : d2) + ((j & 2) != 0 ? -d3 : d3);
 	}
@@ -59,18 +59,18 @@ public class NoiseGeneratorPerlin extends NoiseGenerator {
 			int j3 = 0;
 			double d12 = 1.0D / d6;
 			for (int i4 = 0; i4 < i; i4++) {
-				double d14 = d + (double) i4 * d3 + xCoord;
+				double d14 = d + i4 * d3 + xCoord;
 				int j4 = (int) d14;
-				if (d14 < (double) j4) {
+				if (d14 < j4) {
 					j4--;
 				}
 				int k4 = j4 & 0xff;
 				d14 -= j4;
 				double d17 = d14 * d14 * d14 * (d14 * (d14 * 6D - 15D) + 10D);
 				for (int l4 = 0; l4 < k; l4++) {
-					double d19 = d2 + (double) l4 * d5 + zCoord;
+					double d19 = d2 + l4 * d5 + zCoord;
 					int j5 = (int) d19;
-					if (d19 < (double) j5) {
+					if (d19 < j5) {
 						j5--;
 					}
 					int l5 = j5 & 0xff;
@@ -103,27 +103,27 @@ public class NoiseGeneratorPerlin extends NoiseGenerator {
 		double d16 = 0.0D;
 		double d18 = 0.0D;
 		for (int i5 = 0; i5 < i; i5++) {
-			double d20 = d + (double) i5 * d3 + xCoord;
+			double d20 = d + i5 * d3 + xCoord;
 			int k5 = (int) d20;
-			if (d20 < (double) k5) {
+			if (d20 < k5) {
 				k5--;
 			}
 			int i6 = k5 & 0xff;
 			d20 -= k5;
 			double d22 = d20 * d20 * d20 * (d20 * (d20 * 6D - 15D) + 10D);
 			for (int j6 = 0; j6 < k; j6++) {
-				double d24 = d2 + (double) j6 * d5 + zCoord;
+				double d24 = d2 + j6 * d5 + zCoord;
 				int k6 = (int) d24;
-				if (d24 < (double) k6) {
+				if (d24 < k6) {
 					k6--;
 				}
 				int l6 = k6 & 0xff;
 				d24 -= k6;
 				double d25 = d24 * d24 * d24 * (d24 * (d24 * 6D - 15D) + 10D);
 				for (int i7 = 0; i7 < j; i7++) {
-					double d26 = d1 + (double) i7 * d4 + yCoord;
+					double d26 = d1 + i7 * d4 + yCoord;
 					int j7 = (int) d26;
-					if (d26 < (double) j7) {
+					if (d26 < j7) {
 						j7--;
 					}
 					int k7 = j7 & 0xff;
